Rename P2P tests and move generators into correct package
