+++ core-test/README.md
@@ -10,10 +10,10 @@ chris@chris:~/dev/bitcoins-core$ bloop test coreTest
 
 ## Coverage
 
-To produce a report that quantifies how much of our code is covered by tests: 
+To produce a report that quantifies how much of our code is covered by tests:
 
 ```bash
-sbt 
+sbt
 > coverage
 > coreTest/test
 > core/coverageReport
@@ -21,5 +21,5 @@ sbt
 
 This generates three different reports: Cobertura, XML and HTML formats.
 See the output of your sbt shell to find the location of them.
-Open up the HTML file in your browser. You'll now see code coverage 
-of all files in `core` project. 
+Open up the HTML file in your browser. You'll now see code coverage
+of all files in `core` project.
