+++ core-test/src/test/scala/org/bitcoins/core/p2p/FilterAddMessageTest.scala
@@ -1,12 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.ControlMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.ControlMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-/**
-  * Created by chris on 8/26/16.
-  */
-class FilterAddMessageSpec extends BitcoinSUnitTest {
+class FilterAddMessageTest extends BitcoinSUnitTest {
 
   it must "have serialization symmetry" in {
     forAll(ControlMessageGenerator.filterAddMessage) { filterAddMsg =>
