+++ core-test/src/test/scala/org/bitcoins/core/p2p/GetHeadersMessageTest.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.DataMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.DataMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-class GetHeadersMessageSpec extends BitcoinSUnitTest {
+class GetHeadersMessageTest extends BitcoinSUnitTest {
 
   it must "have serialization symmetry" in {
     forAll(DataMessageGenerator.getHeaderMessages) { headerMsg =>
