+++ core-test/src/test/scala/org/bitcoins/core/p2p/MerkleBlockMessageTest.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.DataMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.DataMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-class MerkleBlockMessageSpec extends BitcoinSUnitTest {
+class MerkleBlockMessageTest extends BitcoinSUnitTest {
   it must "have serialization symmetry" in {
     forAll(DataMessageGenerator.merkleBlockMessage) { merkleBlockMsg =>
       assert(MerkleBlockMessage(merkleBlockMsg.hex) == merkleBlockMsg)
