+++ core-test/src/test/scala/org/bitcoins/core/p2p/PingMessageTest.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.ControlMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.ControlMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-class PingMessageSpec extends BitcoinSUnitTest {
+class PingMessageTest extends BitcoinSUnitTest {
 
   it must "have symmetry serialization" in {
     forAll(ControlMessageGenerator.pingMessage) { pingMessage =>
