+++ core-test/src/test/scala/org/bitcoins/core/p2p/RejectMessageTest.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.ControlMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.ControlMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-class RejectMessageSpec extends BitcoinSUnitTest {
+class RejectMessageTest extends BitcoinSUnitTest {
 
   it must "have serialization symmetry" in {
     forAll(ControlMessageGenerator.rejectMessage) { rejectMsg =>
