+++ core-test/src/test/scala/org/bitcoins/core/p2p/TransactionMessageTest.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.core.p2p
 
-import org.bitcoins.testkit.gen.DataMessageGenerator
+import org.bitcoins.testkit.core.gen.p2p.DataMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 
-class TransactionMessageSpec extends BitcoinSUnitTest {
+class TransactionMessageTest extends BitcoinSUnitTest {
 
   it must "have serialization symmetry" in {
     forAll(DataMessageGenerator.transactionMessage) { txMsg =>
