--- core-test/src/test/scala/org/bitcoins/core/p2p/VersionMessageSpec.scala
@@ -1,13 +1,0 @@
-package org.bitcoins.core.p2p
-
-import org.bitcoins.testkit.gen.ControlMessageGenerator
-import org.bitcoins.testkit.util.BitcoinSUnitTest
-
-class VersionMessageSpec extends BitcoinSUnitTest {
-
-  it must "have serialization symmetry" in {
-    forAll(ControlMessageGenerator.versionMessage) { versionMessage =>
-      assert(VersionMessage(versionMessage.hex) == versionMessage)
-    }
-  }
-}
