+++ core-test/src/test/scala/org/bitcoins/core/p2p/VersionMessageTest.scala
@@ -4,10 +4,17 @@ import java.net.InetAddress
 
 import org.bitcoins.core.config.MainNet
 import org.bitcoins.core.number.{Int32, UInt64}
+import org.bitcoins.testkit.core.gen.p2p.ControlMessageGenerator
 import org.bitcoins.testkit.util.BitcoinSUnitTest
 import org.joda.time.DateTime
 
 class VersionMessageTest extends BitcoinSUnitTest {
+
+  it must "have serialization symmetry" in {
+    forAll(ControlMessageGenerator.versionMessage) { versionMessage =>
+      assert(VersionMessage(versionMessage.hex) == versionMessage)
+    }
+  }
 
   "VersionMessage" must "create a new version message to be sent to another node on the network" in {
     val versionMessage = VersionMessage(MainNet, InetAddress.getLocalHost)
