+++ core/src/main/scala/org/bitcoins/core/serializers/p2p/RawNetworkIpAddressSerializer.scala
@@ -1,11 +1,10 @@
-package org.bitcoins.core.serializers.p2p.messages
+package org.bitcoins.core.serializers.p2p
 
 import java.net.InetAddress
 
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.serializers.RawBitcoinSerializer
 import org.bitcoins.core.util.{BitcoinSLogger, NumberUtil}
-import org.bitcoins.core.p2p.ServiceIdentifier
 import org.bitcoins.core.p2p._
 import scodec.bits.ByteVector
 
@@ -22,7 +21,7 @@ trait RawNetworkIpAddressSerializer
     val services = ServiceIdentifier(bytes.slice(4, 12))
     val ipBytes = bytes.slice(12, 28)
     val ipAddress = InetAddress.getByAddress(ipBytes.toArray)
-    val port = NumberUtil.toLong(bytes.slice(28, 30)).toInt
+    val port = bytes.slice(28, 30).toInt(signed = false)
     NetworkIpAddress(time, services, ipAddress, port)
   }
 
@@ -30,8 +29,9 @@ trait RawNetworkIpAddressSerializer
     val time = networkIpAddress.time.bytes.reverse
     val services = networkIpAddress.services.bytes
     val ipAddress = NetworkIpAddress.writeAddress(networkIpAddress.address)
-    //uint16s are only 4 hex characters
-    val port = ByteVector.fromShort(networkIpAddress.port.toShort)
+    // uint16s are only 4 hex characters
+    // cannot do fromShort,
+    val port = ByteVector.fromInt(networkIpAddress.port, size = 2)
     time ++ services ++ ipAddress ++ port
   }
