+++ core/src/main/scala/org/bitcoins/core/serializers/p2p/messages/RawAddrMessageSerializer.scala
@@ -2,6 +2,7 @@ package org.bitcoins.core.serializers.p2p.messages
 
 import org.bitcoins.core.protocol.CompactSizeUInt
 import org.bitcoins.core.serializers.{RawBitcoinSerializer, RawSerializerHelper}
+import org.bitcoins.core.serializers.p2p._
 import org.bitcoins.core.p2p._
 import scodec.bits.ByteVector
